feat(LIVE-17141): take currency and lack of accounts in consideration
